Reference

Plugins.FieldSpec — Type

FieldSpec(name, type::Type, constructor::Union{Function, DataType} = type)

Field specification for plugin-assembled types.

Note that every field of an assembled type will be constructed with the same arguments. Possibly The constructor will be called when the system

Plugins.HookList — Type

HookList{TNext, THandler, TPlugin}

Provides fast, inlinable call to the implementations of a specific hook.

You can get a HookList by calling hooklist() directly, or using hooks().

The HookList can be called with arbitrary number of extra arguments. If any of the plugins referenced in the list fails to handle the extra arguments, the call will raise a MethodError

Plugins.ImmutableStruct — Type

struct ImmutableStruct <: TemplateStyle end

Plugin-assembled types marked as ImmutableStruct will be generated as a struct.

Plugins.MutableStruct — Type

struct MutableStruct <: TemplateStyle end

Plugin-assembled types marked as MutableStruct will be generated as a mutable struct.

Plugins.Plugin — Type

abstract type Plugin

Provides default implementations of lifecycle hooks.

Plugins.PluginStack — Type

PluginStack(plugins, hookfns = [])

Manages the plugins loaded into an application.

It provides fast access to the plugins by symbol, e.g. pluginstack[:logger]. Collection methods and iteration interface are implemented.

The pluginstack is created from a list of plugins, and optionally a list of hook functions. If hook functions are provided, the hooks()` function can be called to

Plugins.TemplateStyle — Method

TemplateStyle(::Type) = MutableStruct()

Trait to select the template used for plugin-assembled types

Use MutableStruct (default), ImmutableStruct, or subtype it when you want to create your own template.

#Examples

Assembling immutable structs:

abstract type DebugInfo end
Plugins.TemplateStyle(::Type{DebugInfo}) = Plugins.ImmutableStruct()

Defining your own template (see also typedef ):

struct CustomTemplate <: Plugins.TemplateStyle end
Plugins.TemplateStyle(::Type{State}) = CustomTemplate()

Plugins.customfield — Method

customfield(plugin::Plugin, abstract_type::Type, args...)

Provide field specifications to plugin-assembled types.

Using this lifecycle hook the system can define custom plugin-assembled types (typically structs) based on field specifications provided by plugins. E.g. an error type can be extended with debug information.

A plugin can provide zero or one field to every assembled type. To provide a field, return a FieldSpec.

The assembled type will be a subtype of abstract_type. To allow differently configured systems to run in the same Julia session, new types may be assembled for every instance of the system.

Metaprogramming may make you unhappy

Although plugin-assmebled types are designed to help doing metaprogramming in a controlled fashion, it is usually better to use non-meta solutions instead. E.g. Store plugin state inside the plugin, collect data from multiple plugins using lifecycle hooks, etc.

Plugins.customtype — Function

customtype(stack::PluginStack, typename::Symbol, abstract_type::Type, target_module::Module = Main; unique_name = true)

Assemble a type with fields provided by the plugins in stack.

abstract_type will be the supertype of the assembled type.

If unique_name == true, then typename will be suffixed with a structure-dependent id. The id is generated as a hash of the evaluated expression (with the :TYPE_NAME placeholder used instead of the name), meaning that for the same id will be generated for a given type when the same plugins with the same source code are loaded.

Examples

Assembling a type AppStateImpl <: AppState and parametrizing the app with it.

abstract type AppState end

mutable struct CustomFieldsApp{TCustomState}
    state::TCustomState
    function CustomFieldsApp(plugins, hookfns, stateargs...)
        stack = PluginStack(plugins, hookfns)
        state_type = customtype(stack, :AppStateImpl, AppState)
        return new{state_type}(Base.invokelatest(state_type, stateargs...))
    end
end

The need for `invokelatest`

We need to use invokelatest to instantiate a newly generated type. To use the generated type normally, first you have to allow control flow to go to the top-level scope after the type was generated. See also the docs

Antipattern

Assembling state types is an antipattern, because plugins can have their own state. (This may provide better performance in a few cases though) Assembled types can make your code less readable, use them sparingly!

Plugins.hook_cache — Method

hook_cache(stack::PluginStack, hookfns)
hook_cache(plugins, hookfns)

Create a cache of HookLists for a PluginStack or from lists of plugins and hook functions.

Returns a NamedTuple with an entry for every handler.

Examples

cache = hook_cache([Plugin1(), Plugin2()], [hook1, hook2])
cache.hook1()

Plugins.hooklist — Method

function hooklist(plugins, hookfn)

Create a HookList which allows fast, inlinable call to the merged implementations of hookfn by the given plugins.

A plugin of type TPlugin found in plugins will be referenced in the resulting HookList if there is a method that matches the following signature: hookfn(::TPlugin, ...)

Plugins.hooks — Function

hooks(app)
hooks(pluginstack::PluginStack)

Create or get a hook cache for stack.

The first form can be used when pluginstack is stored in app.plugins (the recommended pattern).

When this function is called first time on a PluginStack, the hooks cache will be created by calling hook_cache(), and stored in pluginstack for quick access later.

Plugins.setup! — Method

setup!(plugin, deps, args...)

Initialize the plugin with the given dependencies and arguments (e.g. shared state).

This lifecycle hook will be called when the application loads a plugin. Plugins.jl does not (yet) helps with this, application developers should do it manually, right after the PluginStack was created, before the hook_cache() call.

Plugins.shutdown! — Method

shutdown!(plugin, sharedstate)

Shut down the plugin.

This lifecycle hook will be called when the application unloads a plugin, e.g. before the application exits. Plugins.jl does not (yet) helps with this, application developers should do it manually.

Plugins.symbol — Method

symbol(plugin)

Return the unique Symbol of this plugin if it exports an API to other plugins.

Plugins.typedef — Function

typedef(templatestyle, spec::TypeSpec)::Expr

Return an expression defining a type.

Implement it for your own template styles. More info in the tests.