Reference

Plugins.Plugin — Type

abstract type Plugin

Provides default implementations of lifecycle hooks.

Plugins.PluginStack — Type

PluginStack

Holds all the plugins loaded into an application.

Implements the iteration interface, and gives access using symbols, e.g. pluginstack[:logger]

Plugins.hook_cache — Method

hook_cache(hookfns, sharedstate)

Create a cache of HookLists from the list of hook functions.

Returns a NamedTuple with an entry for every handler.

Examples

cache = hook_cache([hook1, hook2], app)
cache.hook1()

Plugins.hooklist — Method

function hooklist(plugins, hookfn, sharedstate::TSharedState) where {TSharedState}

Create a HookList which allows fast, inlinable call to the merged implementations of hookfn for TSharedState by the given plugins.

A plugin of type TPlugin found in plugins will be referenced in the resulting HookList if there is a method with the following signature: hookfn(::TPlugin, ::TSharedState, ...)

Plugins.hooks — Function

function hooks(stack::PluginStack, rebuild = true)

Plugins.setup! — Method

setup!(plugin, sharedstate)

Initialize the plugin with the given shared state.

This lifecycle hook should be called when the application loads a plugin. Plugins.jl does not (yet) helps with this, application developers should do it manually, right after the PluginStack was created, before the hook_cache() call.

Plugins.shutdown! — Method

shutdown!(plugin, sharedstate)

Shut down the plugin.

This lifecycle hook should be called when the application unloads a plugin, e.g. before the application exits. Plugins.jl does not (yet) helps with this, application developers should do it manually.

Plugins.symbol — Method

symbol(plugin)

Return the unique Symbol of this plugin if it exports an API to other plugins.

Plugins.HookList — Type

HookList{TNext, THandler, TPlugin, TSharedState}

Provides fast, inlinable call to the implementations of a specific hook.

You can get a HookList by calling hooklist() directly, or using hooks().

The HookList can be called with a ::TSharedState and an arbitrary number of extra arguments. If any of the plugins referenced in the list fails to handle the extra arguments, the call will raise a MethodError